[564] feature: use custom workout title prefix
