+++ src/Api.Service/Validators/SyncValidators.cs
@@ -1,6 +1,7 @@
 ﻿using Api.Contract;
 using Api.Service.Helpers;
 using Common;
+using Common.Dto;
 using Common.Stateful;
 using Microsoft.AspNetCore.Mvc;
