+++ src/Api/Services/BackgroundSyncJob.cs
@@ -1,5 +1,6 @@
 ﻿using Common;
 using Common.Database;
+using Common.Dto;
 using Common.Observe;
 using Common.Service;
 using Common.Stateful;
