+++ src/Common/Constants.cs
@@ -9,6 +9,6 @@
 		public const string WebUIName = "p2g_webui";
 		public const string ClientUIName = "p2g_clientui";
 
-		public const string AppVersion = "4.0.0";
+		public const string AppVersion = "4.1.0";
 	}
 }
