+++ src/Common/Database/SettingsDb.cs
@@ -1,11 +1,11 @@
-﻿using Common.Helpers;
+﻿using Common.Dto;
+using Common.Helpers;
 using Common.Observe;
 using JsonFlatFileDataStore;
 using Prometheus;
 using Serilog;
 using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace Common.Database
