+++ src/Common/Stateful/GarminApiAuthentication.cs
@@ -1,4 +1,5 @@
-﻿using Flurl.Http;
+﻿using Common.Dto;
+using Flurl.Http;
 
 namespace Common.Stateful;
