+++ src/Conversion/FitConverter.cs
@@ -62,7 +62,7 @@ namespace Conversion
 
 			var startTime = GetStartTimeUtc(workout);
 			var endTime = GetEndTimeUtc(workout, workoutSamples);
-			var title = WorkoutHelper.GetTitle(workout);
+			var title = WorkoutHelper.GetTitle(workout, settings.Format);
 			var sport = GetGarminSport(workout);
 			var subSport = GetGarminSubSport(workout);
 			var deviceInfo = await GetDeviceInfoAsync(workout.Fitness_Discipline, settings);
