+++ src/Conversion/FitDecoder.cs
@@ -1,7 +1,6 @@
-﻿using Common;
+﻿using Common.Dto;
 using Common.Observe;
 using Dynastream.Fit;
-using Microsoft.VisualBasic;
 using Serilog;
 using System.IO;
