+++ src/Garmin/GarminUploader.cs
@@ -1,4 +1,5 @@
 ﻿using Common;
+using Common.Dto;
 using Common.Observe;
 using Common.Service;
 using Common.Stateful;
