+++ src/SharedUI/Shared/AppSettingsForm.razor
@@ -25,7 +25,7 @@
 </EditForm>
 
 @code {
-	private Common.App appSettings;
+	private App appSettings;
 
 	public AppSettingsForm()
 	{
