+++ src/UnitTests/Common/Database/DbMigrationsTests.cs
@@ -2,6 +2,7 @@
 
 using Common;
 using Common.Database;
+using Common.Dto;
 using Common.Dto.P2G;
 using Moq;
 using Moq.AutoMock;
