+++ src/UnitTests/Conversion/JsonConverterTests.cs
@@ -1,4 +1,5 @@
 ﻿using Common;
+using Common.Dto;
 using Common.Service;
 using Conversion;
 using FluentAssertions;
