+++ src/UnitTests/Peloton/ApiClientTests.cs
@@ -1,4 +1,5 @@
 ﻿using Common;
+using Common.Dto;
 using Common.Service;
 using Common.Stateful;
 using Flurl.Http.Testing;
